+++ tests/integration/harvest/test_compare.py
@@ -1,5 +1,5 @@
 from harvester.harvest import HarvestSource
-from harvester.utils.utils import dataset_to_hash, sort_dataset
+from harvester.utils.general_utils import dataset_to_hash, sort_dataset
 
 
 class TestCompare:
