+++ tests/unit/test_utils.py
@@ -1,7 +1,7 @@
 import pytest
 
 from harvester.utils.ckan_utils import munge_tag, munge_title_to_name
-from harvester.utils.utils import parse_args
+from harvester.utils.general_utils import parse_args
 
 # these tests are copied from
 # https://github.com/ckan/ckan/blob/master/ckan/tests/lib/test_munge.py
